+++ .gitignore
@@ -1,1 +1,2 @@
 grade_data/csvs/*
+*.pyc
