+++ requirements.txt
@@ -1,3 +1,3 @@
-streamlit>=1.35.0
+streamlit>=1.36.0
 plotly>=5.22.0
 pandas>=2.0.0
